backend/requirements.txt
# FastAPI Core
fastapi[all]>=0.115.0
uvicorn>=0.30.0

# Database & Cache
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
redis>=5.0.0

# Trading & Data Science
pandas>=2.2.0
numpy>=2.1.0
ccxt>=4.4.0
python-binance>=1.0.19
scipy>=1.14.0

# Utils & Security
python-dotenv>=1.0.0
pydantic-settings>=2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests>=2.32.0
loguru>=0.7.0

# Telegram
python-telegram-bot>=21.0
